% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMCA.R
\docType{package}
\name{PPMCA}
\alias{PPMCA}
\title{PPMCA: A Shiny web application to analyze species data in spatial and temporal scales and for changepoint analysis}
\description{
The PPMCA package runs a Shiny web application that allows to
visualize spatial and temporal species data, estimate significant factors and identify the significant changes in abundance level.
The application allows to fit Bayesian zero inflation joint models to obtain significant estimates and their uncertainty
by using R-INLA, and to detect changepoints by using four different Bayesian changepoint techniques.
The application allows user interaction and creates interactive visualizations
such as maps supporting padding and zooming and tables that allow for filtering.
}
