% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMCA.R
\docType{package}
\name{PPMCA}
\alias{PPMCA}
\alias{PPMCA-package}
\title{PPMCA: A Shiny web application for Persistence Probability Models and Changepoint Analysis}
\description{
Users can use this app to identify the significant explanatory variables and
the trends of the species during a span of years. The app would use truncated count models
with log linear regression and zero inflation probability models with logistic regression.
This app enables users to identify the significant changes in abundance level using Bayesian changepoint method. Species persistence models in both spatial and temporal scales are used to identify the significant
predictors for species count data while considering zero inflation models and detect significant changes in
species abundance level using changepoint analysis.
}
